requirements.txt
openai
langchain
streamlit
python-dotenv
pypdf

#for installation of local packages
-e .